+++ src/Collector/ExceptionCollector.php
@@ -1,17 +1,16 @@
 <?php
+
 /**
- * Zend Developer Tools for Zend Framework (http://framework.zend.com/)
- *
- * @link       http://github.com/zendframework/ZendDeveloperTools for the canonical source repository
- * @copyright  Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-developer-tools for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-developer-tools/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-developer-tools/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendDeveloperTools\Collector;
+namespace Laminas\DeveloperTools\Collector;
 
-use Zend\Mvc\MvcEvent;
-use Zend\Mvc\Application;
-use ZendDeveloperTools\Exception\SerializableException;
+use Laminas\DeveloperTools\Exception\SerializableException;
+use Laminas\Mvc\Application;
+use Laminas\Mvc\MvcEvent;
 
 /**
  * Exception Data Collector.