+++ src/Listener/EventLoggingListenerAggregate.php
@@ -1,19 +1,18 @@
 <?php
+
 /**
- * Zend Developer Tools for Zend Framework (http://framework.zend.com/)
- *
- * @link http://github.com/zendframework/ZendDeveloperTools for the canonical source repository
- * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-developer-tools for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-developer-tools/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-developer-tools/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendDeveloperTools\Listener;
+namespace Laminas\DeveloperTools\Listener;
 
-use Zend\EventManager\Event;
-use Zend\EventManager\SharedEventManagerInterface;
-use ZendDeveloperTools\Collector\CollectorInterface;
-use ZendDeveloperTools\Collector\EventCollectorInterface;
-use ZendDeveloperTools\Profiler;
+use Laminas\DeveloperTools\Collector\CollectorInterface;
+use Laminas\DeveloperTools\Collector\EventCollectorInterface;
+use Laminas\DeveloperTools\Profiler;
+use Laminas\EventManager\Event;
+use Laminas\EventManager\SharedEventManagerInterface;
 
 /**
  * Listens to defined events to allow event-level collection of statistics.