+++ src/Listener/ProfilerListener.php
@@ -1,22 +1,21 @@
 <?php
+
 /**
- * Zend Developer Tools for Zend Framework (http://framework.zend.com/)
- *
- * @link       http://github.com/zendframework/ZendDeveloperTools for the canonical source repository
- * @copyright  Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-developer-tools for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-developer-tools/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-developer-tools/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendDeveloperTools\Listener;
+namespace Laminas\DeveloperTools\Listener;
 
-use Zend\Mvc\MvcEvent;
-use ZendDeveloperTools\Options;
-use ZendDeveloperTools\Profiler;
-use Zend\EventManager\EventManagerInterface;
-use Zend\EventManager\ListenerAggregateInterface;
-use Zend\EventManager\ListenerAggregateTrait;
-use Zend\ServiceManager\ServiceLocatorInterface;
-use Zend\ServiceManager\Exception\ServiceNotFoundException;
+use Laminas\DeveloperTools\Options;
+use Laminas\DeveloperTools\Profiler;
+use Laminas\EventManager\EventManagerInterface;
+use Laminas\EventManager\ListenerAggregateInterface;
+use Laminas\EventManager\ListenerAggregateTrait;
+use Laminas\Mvc\MvcEvent;
+use Laminas\ServiceManager\Exception\ServiceNotFoundException;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 /**
  * Profiler Listener
@@ -72,8 +71,8 @@ class ProfilerListener implements ListenerAggregateInterface
     {
         $strict     = $this->options->isStrict();
         $collectors = $this->options->getCollectors();
-        $report     = $this->serviceLocator->get('ZendDeveloperTools\Report');
-        $profiler   = $this->serviceLocator->get('ZendDeveloperTools\Profiler');
+        $report     = $this->serviceLocator->get('Laminas\DeveloperTools\Report');
+        $profiler   = $this->serviceLocator->get('Laminas\DeveloperTools\Profiler');
 
         $profiler->setErrorMode($strict);