+++ src/Module.php
@@ -1,23 +1,22 @@
 <?php
+
 /**
- * Zend Developer Tools for Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/ZendDeveloperTools for the canonical source repository
- * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-developer-tools for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-developer-tools/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-developer-tools/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendDeveloperTools;
+namespace Laminas\DeveloperTools;
 
-use Zend\EventManager\EventInterface;
-use Zend\ModuleManager\ModuleEvent;
-use Zend\ModuleManager\ModuleManagerInterface;
-use Zend\ModuleManager\Feature\InitProviderInterface;
-use Zend\ModuleManager\Feature\ConfigProviderInterface;
-use Zend\ModuleManager\Feature\ServiceProviderInterface;
-use Zend\ModuleManager\Feature\BootstrapListenerInterface;
-use Zend\ModuleManager\Feature\ViewHelperProviderInterface;
 use BjyProfiler\Db\Adapter\ProfilingAdapter;
+use Laminas\EventManager\EventInterface;
+use Laminas\ModuleManager\Feature\BootstrapListenerInterface;
+use Laminas\ModuleManager\Feature\ConfigProviderInterface;
+use Laminas\ModuleManager\Feature\InitProviderInterface;
+use Laminas\ModuleManager\Feature\ServiceProviderInterface;
+use Laminas\ModuleManager\Feature\ViewHelperProviderInterface;
+use Laminas\ModuleManager\ModuleEvent;
+use Laminas\ModuleManager\ModuleManagerInterface;
 
 class Module implements
     InitProviderInterface,
@@ -57,15 +56,15 @@ class Module implements
         $eventManager  = $event->getTarget()->getEventManager();
         $configuration = $event->getConfigListener()->getMergedConfig(false);
 
-        if (isset($configuration['zenddevelopertools']['profiler']['enabled'])
-            && $configuration['zenddevelopertools']['profiler']['enabled'] === true
+        if (isset($configuration['laminas-developer-tools']['profiler']['enabled'])
+            && $configuration['laminas-developer-tools']['profiler']['enabled'] === true
         ) {
             $eventManager->trigger(ProfilerEvent::EVENT_PROFILER_INIT, $event);
         }
     }
 
     /**
-     * Zend\Mvc\MvcEvent::EVENT_BOOTSTRAP event callback
+     * Laminas\Mvc\MvcEvent::EVENT_BOOTSTRAP event callback
      *
      * @param  EventInterface $event
      * @throws Exception\InvalidOptionException
@@ -82,16 +81,16 @@ class Module implements
         $sem = $em->getSharedManager();
         $sm  = $app->getServiceManager();
 
-        $options = $sm->get('ZendDeveloperTools\Config');
+        $options = $sm->get('Laminas\DeveloperTools\Config');
 
         if (!$options->isToolbarEnabled()) {
             return;
         }
 
-        $report = $sm->get('ZendDeveloperTools\Report');
+        $report = $sm->get('Laminas\DeveloperTools\Report');
 
         if ($options->canFlushEarly()) {
-            $flushListener = $sm->get('ZendDeveloperTools\FlushListener');
+            $flushListener = $sm->get('Laminas\DeveloperTools\FlushListener');
             $flushListener->attach($em);
         }
 
@@ -100,15 +99,15 @@ class Module implements
         }
 
         if ($options->eventCollectionEnabled()) {
-            $eventLoggingListener = $sm->get('ZendDeveloperTools\EventLoggingListenerAggregate');
+            $eventLoggingListener = $sm->get('Laminas\DeveloperTools\EventLoggingListenerAggregate');
             $eventLoggingListener->attachShared($sem);
         }
 
-        $profilerListener = $sm->get('ZendDeveloperTools\ProfilerListener');
+        $profilerListener = $sm->get('Laminas\DeveloperTools\ProfilerListener');
         $profilerListener->attach($em);
 
         if ($options->isToolbarEnabled()) {
-            $toolbarListener = $sm->get('ZendDeveloperTools\ToolbarListener');
+            $toolbarListener = $sm->get('Laminas\DeveloperTools\ToolbarListener');
             $toolbarListener->attach($em);
         }
 
@@ -128,10 +127,16 @@ class Module implements
     public function getViewHelperConfig()
     {
         return [
+            // Legacy Zend Framework aliases
+            'aliases' => [
+                'ZendDeveloperToolsTime' => 'LaminasDeveloperToolsTime',
+                'ZendDeveloperToolsMemory' => 'LaminasDeveloperToolsMemory',
+                'ZendDeveloperToolsDetailArray' => 'LaminasDeveloperToolsDetailArray',
+            ],
             'invokables' => [
-                'ZendDeveloperToolsTime'        => 'ZendDeveloperTools\View\Helper\Time',
-                'ZendDeveloperToolsMemory'      => 'ZendDeveloperTools\View\Helper\Memory',
-                'ZendDeveloperToolsDetailArray' => 'ZendDeveloperTools\View\Helper\DetailArray',
+                'LaminasDeveloperToolsTime'        => 'Laminas\DeveloperTools\View\Helper\Time',
+                'LaminasDeveloperToolsMemory'      => 'Laminas\DeveloperTools\View\Helper\Memory',
+                'LaminasDeveloperToolsDetailArray' => 'Laminas\DeveloperTools\View\Helper\DetailArray',
             ],
         ];
     }
@@ -143,81 +148,102 @@ class Module implements
     {
         return [
             'aliases' => [
-                'ZendDeveloperTools\ReportInterface' => 'ZendDeveloperTools\Report',
+                'Laminas\DeveloperTools\ReportInterface' => 'Laminas\DeveloperTools\Report',
+
+                // Legacy Zend Framework aliases
+                'ZendDeveloperTools\ReportInterface' => 'Laminas\DeveloperTools\ReportInterface',
+                'ZendDeveloperTools\Report' => 'Laminas\DeveloperTools\Report',
+                'ZendDeveloperTools\EventCollector' => 'Laminas\DeveloperTools\EventCollector',
+                'ZendDeveloperTools\ExceptionCollector' => 'Laminas\DeveloperTools\ExceptionCollector',
+                'ZendDeveloperTools\RouteCollector' => 'Laminas\DeveloperTools\RouteCollector',
+                'ZendDeveloperTools\RequestCollector' => 'Laminas\DeveloperTools\RequestCollector',
+                'ZendDeveloperTools\ConfigCollector' => 'Laminas\DeveloperTools\ConfigCollector',
+                'ZendDeveloperTools\MailCollector' => 'Laminas\DeveloperTools\MailCollector',
+                'ZendDeveloperTools\MemoryCollector' => 'Laminas\DeveloperTools\MemoryCollector',
+                'ZendDeveloperTools\TimeCollector' => 'Laminas\DeveloperTools\TimeCollector',
+                'ZendDeveloperTools\FlushListener' => 'Laminas\DeveloperTools\FlushListener',
+                'ZendDeveloperTools\Profiler' => 'Laminas\DeveloperTools\Profiler',
+                'ZendDeveloperTools\Config' => 'Laminas\DeveloperTools\Config',
+                'ZendDeveloperTools\Event' => 'Laminas\DeveloperTools\Event',
+                'ZendDeveloperTools\StorageListener' => 'Laminas\DeveloperTools\StorageListener',
+                'ZendDeveloperTools\ToolbarListener' => 'Laminas\DeveloperTools\ToolbarListener',
+                'ZendDeveloperTools\ProfilerListener' => 'Laminas\DeveloperTools\ProfilerListener',
+                'ZendDeveloperTools\EventLoggingListenerAggregate' => 'Laminas\DeveloperTools\EventLoggingListenerAggregate',
+                'ZendDeveloperTools\DbCollector' => 'Laminas\DeveloperTools\DbCollector',
             ],
             'invokables' => [
-                'ZendDeveloperTools\Report'             => 'ZendDeveloperTools\Report',
-                'ZendDeveloperTools\EventCollector'     => 'ZendDeveloperTools\Collector\EventCollector',
-                'ZendDeveloperTools\ExceptionCollector' => 'ZendDeveloperTools\Collector\ExceptionCollector',
-                'ZendDeveloperTools\RouteCollector'     => 'ZendDeveloperTools\Collector\RouteCollector',
-                'ZendDeveloperTools\RequestCollector'   => 'ZendDeveloperTools\Collector\RequestCollector',
-                'ZendDeveloperTools\ConfigCollector'    => 'ZendDeveloperTools\Collector\ConfigCollector',
-                'ZendDeveloperTools\MailCollector'      => 'ZendDeveloperTools\Collector\MailCollector',
-                'ZendDeveloperTools\MemoryCollector'    => 'ZendDeveloperTools\Collector\MemoryCollector',
-                'ZendDeveloperTools\TimeCollector'      => 'ZendDeveloperTools\Collector\TimeCollector',
-                'ZendDeveloperTools\FlushListener'      => 'ZendDeveloperTools\Listener\FlushListener',
+                'Laminas\DeveloperTools\Report'             => 'Laminas\DeveloperTools\Report',
+                'Laminas\DeveloperTools\EventCollector'     => 'Laminas\DeveloperTools\Collector\EventCollector',
+                'Laminas\DeveloperTools\ExceptionCollector' => 'Laminas\DeveloperTools\Collector\ExceptionCollector',
+                'Laminas\DeveloperTools\RouteCollector'     => 'Laminas\DeveloperTools\Collector\RouteCollector',
+                'Laminas\DeveloperTools\RequestCollector'   => 'Laminas\DeveloperTools\Collector\RequestCollector',
+                'Laminas\DeveloperTools\ConfigCollector'    => 'Laminas\DeveloperTools\Collector\ConfigCollector',
+                'Laminas\DeveloperTools\MailCollector'      => 'Laminas\DeveloperTools\Collector\MailCollector',
+                'Laminas\DeveloperTools\MemoryCollector'    => 'Laminas\DeveloperTools\Collector\MemoryCollector',
+                'Laminas\DeveloperTools\TimeCollector'      => 'Laminas\DeveloperTools\Collector\TimeCollector',
+                'Laminas\DeveloperTools\FlushListener'      => 'Laminas\DeveloperTools\Listener\FlushListener',
             ],
             'factories' => [
-                'ZendDeveloperTools\Profiler' => function ($sm) {
-                    $a = new Profiler($sm->get('ZendDeveloperTools\Report'));
-                    $a->setEvent($sm->get('ZendDeveloperTools\Event'));
+                'Laminas\DeveloperTools\Profiler' => function ($sm) {
+                    $a = new Profiler($sm->get('Laminas\DeveloperTools\Report'));
+                    $a->setEvent($sm->get('Laminas\DeveloperTools\Event'));
                     return $a;
                 },
-                'ZendDeveloperTools\Config' => function ($sm) {
+                'Laminas\DeveloperTools\Config' => function ($sm) {
                     $config = $sm->get('Configuration');
-                    $config = isset($config['zenddevelopertools']) ? $config['zenddevelopertools'] : null;
+                    $config = isset($config['laminas-developer-tools']) ? $config['laminas-developer-tools'] : null;
 
-                    return new Options($config, $sm->get('ZendDeveloperTools\Report'));
+                    return new Options($config, $sm->get('Laminas\DeveloperTools\Report'));
                 },
-                'ZendDeveloperTools\Event' => function ($sm) {
+                'Laminas\DeveloperTools\Event' => function ($sm) {
                     $event = new ProfilerEvent();
-                    $event->setReport($sm->get('ZendDeveloperTools\Report'));
+                    $event->setReport($sm->get('Laminas\DeveloperTools\Report'));
                     $event->setApplication($sm->get('Application'));
 
                     return $event;
                 },
-                'ZendDeveloperTools\StorageListener' => function ($sm) {
+                'Laminas\DeveloperTools\StorageListener' => function ($sm) {
                     return new Listener\StorageListener($sm);
                 },
-                'ZendDeveloperTools\ToolbarListener' => function ($sm) {
+                'Laminas\DeveloperTools\ToolbarListener' => function ($sm) {
                     return new Listener\ToolbarListener(
                         $sm->get('ViewRenderer'),
-                        $sm->get('ZendDeveloperTools\Config')
+                        $sm->get('Laminas\DeveloperTools\Config')
                     );
                 },
-                'ZendDeveloperTools\ProfilerListener' => function ($sm) {
-                    return new Listener\ProfilerListener($sm, $sm->get('ZendDeveloperTools\Config'));
+                'Laminas\DeveloperTools\ProfilerListener' => function ($sm) {
+                    return new Listener\ProfilerListener($sm, $sm->get('Laminas\DeveloperTools\Config'));
                 },
-                'ZendDeveloperTools\EventLoggingListenerAggregate' => function ($sm) {
-                    $config = $sm->get('ZendDeveloperTools\Config');
+                'Laminas\DeveloperTools\EventLoggingListenerAggregate' => function ($sm) {
+                    $config = $sm->get('Laminas\DeveloperTools\Config');
 
                     return new Listener\EventLoggingListenerAggregate(
                         array_map([$sm, 'get'], $config->getEventCollectors()),
                         $config->getEventIdentifiers()
                     );
                 },
-                'ZendDeveloperTools\DbCollector' => function ($sm) {
+                'Laminas\DeveloperTools\DbCollector' => function ($sm) {
                     $p  = false;
                     $db = new Collector\DbCollector();
 
-                    if ($sm->has('Zend\Db\Adapter\Adapter')) {
-                        $adapter = $sm->get('Zend\Db\Adapter\Adapter');
+                    if ($sm->has('Laminas\Db\Adapter\Adapter')) {
+                        $adapter = $sm->get('Laminas\Db\Adapter\Adapter');
                         if ($adapter instanceof ProfilingAdapter) {
                             $p = true;
                             $db->setProfiler($adapter->getProfiler());
                         }
                     }
 
-                    if (! $p && $sm->has('Zend\Db\Adapter\AdapterInterface')) {
-                        $adapter = $sm->get('Zend\Db\Adapter\AdapterInterface');
+                    if (! $p && $sm->has('Laminas\Db\Adapter\AdapterInterface')) {
+                        $adapter = $sm->get('Laminas\Db\Adapter\AdapterInterface');
                         if ($adapter instanceof ProfilingAdapter) {
                             $p = true;
                             $db->setProfiler($adapter->getProfiler());
                         }
                     }
 
-                    if (! $p && $sm->has('Zend\Db\Adapter\ProfilingAdapter')) {
-                        $adapter = $sm->get('Zend\Db\Adapter\ProfilingAdapter');
+                    if (! $p && $sm->has('Laminas\Db\Adapter\ProfilingAdapter')) {
+                        $adapter = $sm->get('Laminas\Db\Adapter\ProfilingAdapter');
                         if ($adapter instanceof ProfilingAdapter) {
                             $db->setProfiler($adapter->getProfiler());
                         }