+++ src/View/Helper/DetailArray.php
@@ -1,28 +1,18 @@
 <?php
+
 /**
- * ZendDeveloperTools
- *
- * LICENSE
- *
- * This source file is subject to the new BSD license that is bundled
- * with this package in the file LICENSE.txt.
- * It is also available through the world-wide-web at this URL:
- * http://framework.zend.com/license/new-bsd
- * If you did not receive a copy of the license and are unable to
- * obtain it through the world-wide-web, please send an email
- * to license@zend.com so we can send you a copy immediately.
- *
- * @copyright  Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd     New BSD License
+ * @see       https://github.com/laminas/laminas-developer-tools for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-developer-tools/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-developer-tools/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendDeveloperTools\View\Helper;
+namespace Laminas\DeveloperTools\View\Helper;
 
-use Zend\View\Helper\AbstractHelper;
+use Laminas\View\Helper\AbstractHelper;
 
 /**
- * @copyright  Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
- * @license    http://framework.zend.com/license/new-bsd     New BSD License
+ * @copyright  Copyright (c) 2005-2014 Laminas (https://www.zend.com)
+ * @license    https://getlaminas.org/license/new-bsd     New BSD License
  */
 class DetailArray extends AbstractHelper
 {
@@ -38,11 +28,11 @@ class DetailArray extends AbstractHelper
     {
         $r   = [];
 
-        $r[] = '<span class="zdt-toolbar-info';
-        $r[] = ($redundant) ? ' zdt-toolbar-info-redundant' : '';
+        $r[] = '<span class="laminas-toolbar-info';
+        $r[] = ($redundant) ? ' laminas-toolbar-info-redundant' : '';
         $r[] = '">';
 
-        $r[] = '<span class="zdt-detail-label">';
+        $r[] = '<span class="laminas-detail-label">';
         $r[] = $label;
         $r[] = '</span>';
 
@@ -52,15 +42,15 @@ class DetailArray extends AbstractHelper
 
         foreach ($details as $entry) {
             if ($newLine === true) {
-                $r[] = '</span><span class="zdt-toolbar-info';
-                $r[] = ($redundant) ? ' zdt-toolbar-info-redundant' : '';
+                $r[] = '</span><span class="laminas-toolbar-info';
+                $r[] = ($redundant) ? ' laminas-toolbar-info-redundant' : '';
                 $r[] = '">';
             }
 
-            $r[] = sprintf('<span class="zdt-detail-value%s">%s</span>', $extraCss, $entry);
+            $r[] = sprintf('<span class="laminas-detail-value%s">%s</span>', $extraCss, $entry);
 
             $newLine  = true;
-            $extraCss = ' zdt-detail-extra-value';
+            $extraCss = ' laminas-detail-extra-value';
         }
 
         $r[] = '</span>';